MiniMax-H3: move the shared media helpers out of the cache CLI into minimax_h3/media.py
The minimax_h3 package (generation_inputs), the generation script and the
trainer all imported PyAVH3MediaDecoder and fingerprint_file from
minimax_h3_cache_latents.py, and the text-encoder cache script pulled two
private helpers (_adapt_canvas, _resize_frames) from the same entry point.
A library module depending on a CLI script is the wrong direction (every
other architecture's cache scripts import *from* their package), it dragged
the whole dataset/config stack into anyone importing generation_inputs, and it
left near-copies of the same helpers on both sides of the boundary.

minimax_h3/media.py now owns the reference decoding and pixel helpers:
- H3MediaDecoder / PyAVH3MediaDecoder, adapt_reference_canvas,
  round_to_canvas_multiple, resize_frames (with the released canvas
  constants CANVAS_MULTIPLE / BASE_SHORT_EDGE / MAX_PIXELS)
- prepare_pixels (the cache version, a superset of the generation copy:
  accepts ndarray and drops alpha) and module_device_dtype (was duplicated
  as _model_device_dtype)
- fingerprint_file; fingerprint_checkpoint moves next to
  resolve_safetensors_files in minimax_h3/checkpoint.py
- TEXT_VISUAL_FPS / TEXT_VISUAL_FRAME_STRIDE: the 2 fps text-visual clock
  that was a named constant in one script and the literal 12 / 2.0 in
  generation_inputs and text_encoder

The cache scripts, the trainer and the generation script import from the
package. The text-encoder cache script still imports the dataset-identity
helpers from the latent cache script; that coupling is the next cleanup.

A new test asserts the import direction (no minimax_h3/ module, trainer or
generation script imports a cache script) so it cannot regress.
